backend/auth_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
requests-oauthlib==1.3.1
pydantic==2.4.2
python-dotenv==1.0.0
